Create MarsPopModel
You are a student researcher on the Mars rover UIU team studying a unique species of organisms found exclusively on planet Mars. These organisms have the ability to reproduce themselves once every day. Assuming an initial population of one, what is the estimated total number of these organisms at the end of a given month, if research began on the first day of that month?